print-canon: Drop use of hex arguments to canon_cmd()
What the hell is ESC28? And what the heck does 0x65 do?
Instead, use strings to describe the command preamble. This is
consistent with hexdump, wireshark, and pixma_parse output formats,
and is actually half-readable.
Drop the 'ESC' argument since every command is prefixed by ESC (0x1b).
In the new form, it's infinitely easier to search for commands.
For example, to find occurences of ESC (P:
	Old grep regex: 'ESC28[[:space:]]*,[[:space:]]*0x65'
	New grep regex: '(P'

As a bonus, we no longer have to memorize the ASCII table in
hexadecimal form, as ASCII characters are represented as strings.