print-canon: Add support for SetTime setup command
Some printers use the time information in ESC [K to determine if it's
time to run a cleaning cycle. Lack of this causes these printers to
run a cleaning cycle, leaving gutenprint users with empty ink tanks,
and more junk in the waste tanks.

This was tested on a Pro-100, and without SetTime it appeared that the
printer wanted to dump ink before each print job. With SetTime, this
is no longer the case. I suspect this could be enabled for all Canon
printers, but it's implemented as a new CANON_CAP in order to maintain
behavior of existing printers.