print-canon: Add support for ESC (N data format
ESC (N is very similar to ESC (F, but it supports a 32-bit length
field instead of 16-bit. The limited length might become an issue on
wide papers printed at 1200 DPI.

Newer printers, such as the Pro-100 understand this new format. Due to
its length limitation, ESC (F is not suitable for all print jobs, and
ESC (N is preferred.